fix: Update chunkSizeInBlocks_ closer to its use
Update of chunkSizeInBlocks_ is necessary in some tests to be closer to
the point where it is used: the call to startNewOperations. It appears
that the acceptation of new operations makes updates to the file size in
a way that makes the value set previously just too old. This commit fix
some tests where the described issue appers,