fix: Update maxfleng with locator's fileLength
Some tests performed when multiple clients are writing and enlarging a
single file showed that maxfleng value was not perfectly updated only
with the information a single client had. Thus, it was necessary to
update it after locking the chunk and that was the one value fileLength
value got by the locator provided.