feat(tests): Disable test_richacl test
Currently, this test fails for unknown reasons. While there is some
debate to be had for on whether to deprecate RichACL support (at least
the interface on Linux, since it was never supported anyway), this test
currently doesn't require our attention since there's no use case for
this that needs to be tested for.