fix(tests): Various fixes for the testing framework
This is a series of 3 commits:

tidy(tests): Remove unnecessary ticks

This messes up some editor syntax highlighting

fix(tests): Fix binary output being outputted

The variable used in the conditional is different from the actual
filename used. This conditional would always succeed since awk would
return 0 if the conditional in awk would not match, which it didn't
since stdout was empty (only stderr was populated with the error
message, but it is disregarded).

fix(tests): Coredumps not being generated

At least 3 years ago, `sudo` had a change of behaviour regarding
coredumps. This fixes that by setting the default for cores as the
system.